# model/gwapp-types.gen.go
// Package model provides primitives to interact the openapi HTTP API.
//
// Code generated by github.com/deepmap/oapi-codegen DO NOT EDIT.
package model

// Error defines model for Error.
type Error struct {

	// Error code
	Code int32 `json:"code"`

	// Error message
	Message string `json:"message"`
}

// Event defines model for Event.
type Event struct {

	// Unique id of the Event
	Id *int64 `json:"id,omitempty"`

	// Type of the event
	Location *string `json:"location,omitempty"`

	// Name of the event
	Name string `json:"name"`

	// Type of the event
	Tag *string `json:"tag,omitempty"`
}

// FindEventsParams defines parameters for FindEvents.
type FindEventsParams struct {

	// tags to filter by
	Tags *[]string `json:"tags,omitempty"`

	// maximum number of results to return
	Limit *int32 `json:"limit,omitempty"`
}

// AddEventJSONBody defines parameters for AddEvent.
type AddEventJSONBody Event

// AddEventRequestBody defines body for AddEvent for application/json ContentType.
type AddEventJSONRequestBody AddEventJSONBody